Further Storage refactoring
